Made the YouTube side of the page resizable
Through some clever trickery, I managed to make the YouTube iFrame
resizable. I used a guide on the website listed in the comment next to
the CSS - basically, the iFrame is set to the absolute size of a parent
wrapper than CAN resize, but will always do so in the 16:9 aspect ratio.